+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - 使用快捷键 <kbd>Shift + Alt + C</kbd> 复制当前文档 markdown 全文至剪贴板 | Use the shortcut keys <kbd>Shift + Alt + C</kbd> to copy the full markdown text of the current document to the clipboard.
 - 使用快捷键 <kbd>Shift + Alt + D</kbd> 删除当前文档全部内容 | Use the shortcut keys <kbd>Shift + Alt + D</kbd> to delete the full content of the current documen.
 - 添加图片元素的阴影 | Adds a shadow of a picture element.
+- 优化列表转表格功能 | Optimized the list to table function.
 
 ## v0.4.0/2022-03-03